+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 - Bump JavaScript SDK from v10.37.0 to v10.38.0 ([#5596](https://github.com/getsentry/sentry-react-native/pull/5596))
   - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#10380)
   - [diff](https://github.com/getsentry/sentry-javascript/compare/10.37.0...10.38.0)
+- Bump Bundler Plugins from v4.8.0 to v4.9.0 ([#5608](https://github.com/getsentry/sentry-react-native/pull/5608))
+  - [changelog](https://github.com/getsentry/sentry-javascript-bundler-plugins/blob/main/CHANGELOG.md#490)
+  - [diff](https://github.com/getsentry/sentry-javascript-bundler-plugins/compare/4.8.0...4.9.0)
 
 ## 8.0.0-alpha.0